+++ src/frontend/vite-project/src/components/views/CategoryView.jsx
@@ -1,10 +1,10 @@
-import { Title } from './Title.jsx'
-import { TransactionPanel } from './TransactionPanel.jsx'
-import { CategoriesDoughnutChart } from './charts/CategoriesDoughnutChart.jsx';
-import { LoadingDots } from './LoadingDots.jsx';
-import { SubcategoriesList } from './SubcategoriesList.jsx';
-import { ActionsMenuAdd } from './ActionsMenuAdd.jsx';
-import { ActionsMenuEditDelete } from './ActionsMenuEditDelete.jsx';
+import { Title } from '../Title.jsx'
+import { TransactionPanel } from '../TransactionPanel.jsx'
+import { CategoriesDoughnutChart } from '../charts/CategoriesDoughnutChart.jsx';
+import { LoadingDots } from '../LoadingDots.jsx';
+import { SubcategoriesList } from '../SubcategoriesList.jsx';
+import { ActionsMenuAdd } from '../ActionsMenuAdd.jsx';
+import { ActionsMenuEditDelete } from '../ActionsMenuEditDelete.jsx';
 import { useLocation } from 'react-router-dom';
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
