+++ src/frontend/vite-project/src/components/views/FormCategoryBudget.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import { Menu, MenuButton, MenuItem, MenuItems } from '@headlessui/react';
 import { ChevronDownIcon } from '@heroicons/react/20/solid';
-import { Title } from './Title.jsx';
+import { Title } from '../Title.jsx';
 import { useNavigate } from 'react-router-dom';
 import { useLocation } from 'react-router-dom';
 import axios from 'axios';
