+++ src/frontend/vite-project/src/components/views/FormExpenseOCR.jsx
@@ -1,12 +1,12 @@
 import { useState, useEffect } from 'react';
 import { Menu, MenuButton, MenuItem, MenuItems } from '@headlessui/react';
 import { ChevronDownIcon } from '@heroicons/react/20/solid';
-import { Title } from './Title.jsx';
+import { Title } from '../Title.jsx';
 import { useNavigate } from 'react-router-dom';
 import { useLocation } from 'react-router-dom';
 import axios from 'axios';
-import { LocationMiniMap } from './LocationMiniMap.jsx';
-import { FindLocationMiniMap } from './FindLocationMiniMap.jsx';
+import { LocationMiniMap } from '../LocationMiniMap.jsx';
+import { FindLocationMiniMap } from '../FindLocationMiniMap.jsx';
 
 export function FormExpenseOCR() {
   const location = useLocation();
