+++ src/frontend/vite-project/src/components/views/FormSaving.jsx
@@ -1,40 +1,57 @@
 import { useState, useEffect } from 'react';
-import { Title } from './Title.jsx';
-
+import { Menu, MenuButton, MenuItem, MenuItems } from '@headlessui/react';
+import { ChevronDownIcon } from '@heroicons/react/20/solid';
+import { Title } from '../Title.jsx';
 import { useNavigate } from 'react-router-dom';
 import { useLocation } from 'react-router-dom';
 import axios from 'axios';
 
-export function FormIncome() {
+export function FormSaving() {
   const location = useLocation();
   const state = location.state;
   console.log(state);
-  
+
   const [amount, setAmount] = useState('');
-  const [name, setDescription] = useState('');
+  const [name, setConcept] = useState('');
+  const [selectedGoal, setSelectedGoal] = useState('');
   const [payment_method, setPaymentMethod] = useState('');
   const [insertDate, setInsertDate] = useState('');
+  const [savingGoals, setSavingGoals] = useState([]);
 
   const navigate = useNavigate();
+  let savingGoals_list = [];
 
   useEffect(() => {
-    console.log('stateIncome:', state);
-    if (state.transaction_id) {
+    axios.get('http://localhost:8000/goals')
+        .then(response => {
+          savingGoals_list = response.data;
+          setSelectedGoal(savingGoals_list.find(goal => goal.name === "Otros"));
+          setInsertDate(new Date().toISOString().split('T')[0]);
+          setSavingGoals(savingGoals_list);
+          })
+          .catch(error => {
+            console.error('Error fetching the goals:', error);
+        });
+
+    if(state.transaction_id) {
       axios.get(`http://localhost:8000/transaction/${state.transaction_id}`)
         .then(response => {
-          const income = response.data;
-          console.log('income:', income);
-          setAmount(income.amount);
-          setDescription(income.name);
-          setInsertDate(new Date(income.insert_date).toISOString().split('T')[0]);
-          setPaymentMethod(income.payment_method);
+          const saving = response.data;
+          setAmount(saving.amount);
+          setConcept(saving.name);
+          setInsertDate(new Date(saving.insert_date).toISOString().split('T')[0]);
+          handleGoalChange(savingGoals_list.find(goal => goal.id === saving.saving_goal_id));
+          setPaymentMethod(saving.payment_method);
         })
         .catch(error => {
-          console.error('Error fetching income data:', error);
+            console.error('Error fetching saving data:', error);
         });
     }
   }, []);
 
+  const handleGoalChange = (saving_goal) => {
+    setSelectedGoal(saving_goal);
+  };
 
   const handlePaymentMethodChange = (type) => {
     setPaymentMethod(type);
@@ -43,48 +60,48 @@ export function FormIncome() {
   const handleSubmit = async (event) => {
     event.preventDefault();
 
-    const newIncome = {
+    const newSaving = {
       amount: parseFloat(amount),
       name,
+      saving_goal_id: selectedGoal.id,
       payment_method,
       user_id: 1,
-      transaction_type: "Income",
+      transaction_type: "Saving",
       insert_date: insertDate ? insertDate : null,
     };
 
-    console.log('newIncome:', newIncome);
-
     if (state.transaction_id) {
       try {
-        await axios.put(`http://localhost:8000/transaction/${state.transaction_id}`, newIncome);
-        console.log('Ingreso actualizado con éxito');
-        navigate('/transactions', { state: { transaction_type: "incomes" } });
+        await axios.put(`http://localhost:8000/transaction/${state.transaction_id}`, newSaving);
+        console.log('Ahorro actualizado con éxito');
+        navigate('/transactions', { state: { transaction_type: "savings" } });
       } catch (error) {
-        console.error('Error al actualizar ingreso:', error.response.statusText);
+        console.error('Error al actualizar ahorro:', error.response.statusText);
       };
 
     }else{
       try {
-        await axios.post('http://localhost:8000/transactions', newIncome);
-        console.log('Ingreso creado con éxito');
-        navigate('/transactions', { state: { transaction_type: "incomes" } });
+        await axios.post('http://localhost:8000/transactions', newSaving);
+        console.log('Ahorro creado con éxito');
+        navigate('/transactions', { state: { transaction_type: "savings" } });
       } catch (error) {
-        console.error('Error al crear ingreso:', error.response.statusText);
+        console.error('Error al crear ahorro:', error.response.statusText);
       };
     };
-  };
-
-  const handleCancel = () => {
-    navigate('/transactions', { state: { transaction_type: "incomes" } });
   }
 
+  const handleCancel = () => {
+    navigate('/transactions', { state: { transaction_type: "savings" } });
+  }
+    
+
   return (
-    <div className="flex min-h-full flex-1 flex-col justify-center px-6 py-12 lg:px-8">
-    { state.transaction_id ? (
-      <Title title="Editar Ingreso" />
-    ) : (
-      <Title title="Añadir Ingreso" />
-    )}
+    <div className="flex min-h-full flex-1 flex-col justify-center px-6 py-12 lg:px-8">  
+      { state.transaction_id ? (
+        <Title title="Editar Ahorro" />
+      ) : (
+        <Title title="Añadir Ahorro" />
+      )}
       <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-lg">
         <form onSubmit={handleSubmit} className="space-y-6">
           <div>
@@ -116,8 +133,9 @@ export function FormIncome() {
                 id="name"
                 name="name"
                 value={name}
-                onChange={(e) => setDescription(e.target.value)}
-                placeholder='Ej. Nómina'
+                onChange={(e) => setConcept(e.target.value)}
+                required
+                placeholder='Ej. Ahorro para el viaje'
                 className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
               />
             </div>
@@ -142,7 +160,35 @@ export function FormIncome() {
           </div>
 
           <div>
-            <label className="block text-sm font-medium leading-6 text-gray-900">Tipo de Ingreso</label>
+            <label htmlFor="selectedGoal" className="block text-sm font-medium leading-6 text-gray-900">
+              Objetivo de Ahorro
+            </label>
+            <div className="mt-2">
+              <Menu as="div" className="relative inline-block text-left w-full">
+                <div>
+                  <MenuButton className="inline-flex w-full justify-between gap-x-1.5 rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50">
+                    {selectedGoal ? selectedGoal.name : 'Seleccionar objetivo'}
+                    <ChevronDownIcon aria-hidden="true" className="-mr-1 h-5 w-5 text-gray-400" />
+                  </MenuButton>
+                </div>
+                <MenuItems className="absolute z-10 mt-2 w-full origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+                  {savingGoals.map((saving_goal) => (
+                    <MenuItem
+                      key={saving_goal.id}
+                      as="button"
+                      className="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 hover:text-gray-900"
+                      onClick={() => handleGoalChange(saving_goal)}
+                    >
+                      {saving_goal.name}
+                    </MenuItem>
+                  ))}
+                </MenuItems>
+              </Menu>
+            </div>
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium leading-6 text-gray-900">Tipo de Ahorro</label>
             <div className="mt-2 flex justify-around">
               <button
                 type="button"
@@ -164,7 +210,7 @@ export function FormIncome() {
           <div>
             <button
               type="submit"
-              className="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
+              className="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 my-2 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
             >
               { state.transaction_id ? (
                 <p>Editar</p>
