+++ src/frontend/vite-project/src/components/views/FormSavingGoal.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import { Menu, MenuButton, MenuItem, MenuItems } from '@headlessui/react';
-import { Title } from './Title.jsx';
+import { Title } from '../Title.jsx';
 import { useNavigate } from 'react-router-dom';
 import { useLocation } from 'react-router-dom';
 import axios from 'axios';
