+++ src/frontend/vite-project/src/components/views/FormSubcategoryBudget.jsx
@@ -1,6 +1,6 @@
 import { Menu, MenuButton, MenuItem, MenuItems } from '@headlessui/react';
 import { ChevronDownIcon } from '@heroicons/react/20/solid';
-import { Title } from './Title.jsx';
+import { Title } from '../Title.jsx';
 import { useNavigate } from 'react-router-dom';
 import { useState, useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
