+++ src/frontend/vite-project/src/components/views/SavingView.jsx
@@ -1,8 +1,8 @@
-import { Title } from './Title.jsx'
-import { TransactionPanel } from './TransactionPanel.jsx'
-import { SavingsLineChart } from './charts/SavingsLineChart.jsx';
-import { LoadingDots } from './LoadingDots.jsx';
-import { ActionsMenuEditDelete } from './ActionsMenuEditDelete.jsx';
+import { Title } from '../Title.jsx'
+import { TransactionPanel } from '../TransactionPanel.jsx'
+import { SavingsLineChart } from '../charts/SavingsLineChart.jsx';
+import { LoadingDots } from '../LoadingDots.jsx';
+import { ActionsMenuEditDelete } from '../ActionsMenuEditDelete.jsx';
 import { useLocation } from 'react-router-dom';
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
